+++ docs/Changelog.md
@@ -7,6 +7,7 @@
   - `ThemeCallback.getBitmap`
 - Update JTS, OkHttp dependencies [#1172](https://github.com/mapsforge/vtm/pull/1172) [#1173](https://github.com/mapsforge/vtm/pull/1173)
 - Remove SLF4J dependency [#1167](https://github.com/mapsforge/vtm/pull/1167) [#1170](https://github.com/mapsforge/vtm/pull/1170)
+- Libtess2 v1.0.1 [#1178](https://github.com/mapsforge/vtm/pull/1178)
 - Disable `vtm-ios` [#1166](https://github.com/mapsforge/vtm/pull/1166)
 - Remove `vtm-theme-comparator` [#1169](https://github.com/mapsforge/vtm/pull/1169)
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.24.0)