+++ docs/Changelog.md
@@ -2,6 +2,7 @@
 
 ## Next version
 
+- Android native libs for 16 KB page size [#1179](https://github.com/mapsforge/vtm/pull/1179)
 - Multi-map `MapFileTileSource.setPriority` [#1176](https://github.com/mapsforge/vtm/pull/1176)
 - Color filter theme resources [#1175](https://github.com/mapsforge/vtm/pull/1175)
   - `ThemeCallback.getBitmap`